split out forum thread code into its own resource
