+++ server/src/resources/forum/thread/model.go
@@ -1,0 +1,1 @@
+package thread
